Move test classes into spec/model
