+++ spec/trackable_spec.rb
@@ -1,38 +1,5 @@
 require "spec_helper"
 
-class Student
-  def initialize(name, age)
-    @name = name
-    @age = age
-  end
-
-  def self.foo; end
-
-  def name
-    @name
-  end
-
-  def age
-    @age
-  end
-
-  def age=(age)
-    @age = age
-  end
-end
-
-class HighSchoolStudent < Student;end
-
-class School
-  def initialize(name)
-    @name = name
-  end
-
-  def name
-    @name
-  end
-end
-
 RSpec.describe TappingDevice::Trackable do
   include described_class
